# docs/swagger/swagger.json
{
  "basePath": "/",
  "definitions": {
    "V1BearerAuthError": {
      "properties": {
        "error": {
          "properties": {
            "message": {
              "example": "unauthenticated",
              "type": "string"
            }
          },
          "required": [
            "message"
          ],
          "type": "object"
        }
      },
      "required": [
        "error"
      ],
      "type": "object"
    },
...
    "V1ListViewEnvelope": {
      "properties": {
        "items": {
          "items": {},
          "type": "array"
        },
        "meta": {}
      },
      "required": [
        "items"
      ],
      "type": "object"
    },
    "V1NotImplementedError": {
      "properties": {
        "error": {
          "properties": {
            "capability": {
              "type": "string"
            },
            "code": {
              "example": "not_implemented",
              "type": "string"
            },
            "implemented": {
              "example": false,
              "type": "boolean"
            },
            "message": {
              "type": "string"
            }
          },
          "required": [
            "code",
            "message",
            "capability",
            "implemented"
          ],
          "type": "object"
        }
      },
      "required": [
        "error"
      ],
      "type": "object"
    },
...
    "V1StandardError": {
      "properties": {
        "error": {
          "properties": {
            "code": {
              "example": "invalid_json",
              "type": "string"
            },
            "message": {
              "type": "string"
            }
          },
          "required": [
            "code",
            "message"
          ],
          "type": "object"
        }
      },
      "required": [
        "error"
      ],
      "type": "object"
    },
...
  },
  "host": "localhost:8080",
  "info": {
    "contact": {
      "name": "AVF Engineering",
      "url": "https://github.com/avf/avf-vending-system"
    },
    "description": "HTTP API for the AVF vending platform (cmd/api). All `/v1/*` routes require `Authorization: Bearer <JWT>` unless noted (health, optional metrics, Swagger when enabled). Bearer middleware failures use minimal JSON `{\"error\":{\"message\":\"...\"}}` without `error.code`. Most `/v1` handlers use `{\"error\":{\"code\":\"...\",\"message\":\"...\"}}` (see writeAPIError). **501** list stubs return `not_implemented` plus `capability` and `implemented:false`. **503** `capability_not_configured` is used when optional wiring (MQTT dispatch, commerce persistence, outbox defaults) is missing. Request tracing: responses echo `X-Request-ID` and `X-Correlation-ID` when middleware is enabled.",
    "license": {
      "name": "License terms apply per your deployment agreement; see repository NOTICE if present."
    },
    "termsOfService": "https://github.com/avf/avf-vending-system/tree/main/avf-vending-api",
    "title": "AVF Vending HTTP API",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/admin/assignments": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns **501** with capability `v1.admin.assignments.list`.",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "501": {
            "description": "List technician assignments (admin)",
            "schema": {
              "$ref": "#/definitions/V1NotImplementedError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List technician assignments (admin)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/commands": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns **501** with capability `v1.admin.commands.list`.",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "501": {
            "description": "List commands (admin)",
            "schema": {
              "$ref": "#/definitions/V1NotImplementedError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List commands (admin)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/machines": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Postgres-backed fleet listing when Fleet service is wired. Returns **501** with `not_implemented` when fleet is nil or misconfigured (`v1.admin.machines.list`). Platform admin must supply tenant scope (organization_id on principal) or receives **400** `tenant_scope_required`.",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "List machines (admin)",
            "schema": {
              "$ref": "#/definitions/V1ListViewEnvelope"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "501": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1NotImplementedError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List machines (admin)",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/v1/admin/ota": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns **501** with capability `v1.admin.ota.list`.",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "501": {
            "description": "List OTA artifacts (admin)",
            "schema": {
              "$ref": "#/definitions/V1NotImplementedError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List OTA artifacts (admin)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/technicians": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Returns **501** with `not_implemented` and capability `v1.admin.technicians.list` until a repository is wired.",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "501": {
            "description": "List technicians (admin)",
            "schema": {
              "$ref": "#/definitions/V1NotImplementedError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List technicians (admin)",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/v1/machines/{machineId}/shadow": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Requires Bearer JWT and RequireMachineURLAccess(machineId). Invalid UUID in path \u2192 **400** from middleware or handler (`invalid_machine_id`). Missing shadow row \u2192 **404** `machine_shadow_not_found`.",
        "parameters": [
          {
            "description": "Machine UUID",
            "in": "path",
            "name": "machineId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "Get machine shadow JSON",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "404": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get machine shadow JSON",
        "tags": [
          "Fleet"
        ]
      }
    },
...
    "/v1/orders": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "**501** with capability `v1.orders.org_list`. Same auth and 403 behavior as payments list.",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "501": {
            "description": "List orders for organization",
            "schema": {
              "$ref": "#/definitions/V1NotImplementedError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List orders for organization",
        "tags": [
          "Commerce"
        ]
      }
    },
    "/v1/payments": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Requires Bearer JWT and RequireOrganizationScope. **501** `not_implemented` with capability `v1.payments.org_list` until implemented. Non-platform callers without org on JWT get **403** from middleware (minimal JSON, no `code`).",
        "produces": [
          "json"
        ],
        "responses": {
          "401": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "403": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1BearerAuthError"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/V1StandardError"
            }
          },
          "501": {
            "description": "List payments for organization",
            "schema": {
              "$ref": "#/definitions/V1NotImplementedError"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List payments for organization",
        "tags": [
          "Commerce"
        ]
      }
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Send: Bearer <JWT>. JWT mode is selected by HTTP_AUTH_MODE (hs256 default, rs256_pem, rs256_jwks). 401/403 responses from this layer use JSON {\"error\":{\"message\":\"...\"}} without error.code.",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0",
...
}